pyproject.toml
[tool.poetry]
name = "FraseDeHoy"
version = "1.0.0"
description = "Bot que uso a veces en mi cuenta de Mastodon para subir frases de mi base de datos de frases usando el hashtag #FraseDeHoy."
authors = ["Luqaska"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
frases = "^0.1.3"
"Mastodon.py" = "^1.5.1"

[tool.poetry.dev-dependencies]
frases = "^0.1.3"
"Mastodon.py" = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"